add starlingx api endpoints
This commits introduces a new series of API endpoints for the StarlingX
API.  The code will be updated to conform to the test requirements of
the Gophercloud project with the end goal of merging each sub directory
over to the master branch piece by piece.

See the following website for more information regarding the StarlingX
project.

        https://www.starlingx.io/

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>