starlingx: restore support for interface network fields
To support older API versions it has been necessary to restore support
for certain network related interfaces fields.  The newer API versions
use the "interface-network" associations, but the older API version use
the fields directly on the Interface objects.

We will hopefully be able to remove these again in the near future when
we discontinue support for the older loads that are requiring this
change.

This was an in-cycle development change to the API not a formal API
change that requires proper backwards compatibility.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>