starlingx: add host utility to check for stable state
This commits adds a utility that allows a caller to check a host for a
"stable" state.  A "stable" host is one that has reached its intended
operational state and is not currently performing a task (i.e., idle).
For example, if a Host is "locked/enabled" then it is clearly
transititioning to the "locked/disabled" state therefore it is not
stable, but a host that is "locked/disabled" is stable unless there is a
pending task.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>